+++ Runtime/SO/DialogueContainerValues.cs
@@ -2,7 +2,6 @@
 
 namespace Fenneig_Dialogue_Editor.Runtime.SO
 {
-    [CreateAssetMenu(menuName = "Dialogue/Dialogue container")]
     public class DialogueContainerValues : ScriptableObject
     {